# Performance improvements for bfs