@alejandromarcu 2025-05-07
@jonbinney , I'll be curious to see if you also see a 4x perf improvement in simple, if you try it lmk