@jonbinney 2025-05-07
@alejandromarcu yep i do!

Before this PR:
```
+---------+-------------------+---------------+---------------+
|  Player | Average Time (ms) | Min Time (ms) | Max Time (ms) |
+---------+-------------------+---------------+---------------+
| simple  |      413.231      |     33.607    |    928.710    |
|  greedy |       3.305       |     1.316     |     6.488     |
+---------+-------------------+---------------+---------------+
```

After:
```
+---------+-------------------+---------------+---------------+
|  Player | Average Time (ms) | Min Time (ms) | Max Time (ms) |
+---------+-------------------+---------------+---------------+
| simple  |       93.223      |     11.271    |    157.391    |
|  greedy |       1.388       |     0.699     |     2.461     |
+---------+-------------------+---------------+---------------+
```