@alejandromarcu 2025-05-07
Great! but the bad news is that to make it to depth 4 for example we would need about 100x improvement in speed. I think it's mainly measuring the distances from the pawns to the target when max_depth = 0 and checking for so many wall placements.